+++ GIAN07/player/loadout/player_loadout.h
@@ -28,8 +28,8 @@ struct PlayerTraits {
   int move_speed;
   int focus_move_speed;
   int hit_radius;
-  uint16_t bomb_duration;
-  uint8_t option_sprite;
+  int bomb_duration;
+  int option_sprite;
   int option_offset;
   int focus_option_offset;
 };
@@ -47,25 +47,25 @@ public:
     return focused ? traits_.focus_move_speed : traits_.move_speed;
   }
   [[nodiscard]] int HitRadius() const { return traits_.hit_radius; }
-  [[nodiscard]] uint16_t BombDuration() const { return traits_.bomb_duration; }
-  [[nodiscard]] uint8_t OptionSprite() const { return traits_.option_sprite; }
+  [[nodiscard]] int BombDuration() const { return traits_.bomb_duration; }
+  [[nodiscard]] int OptionSprite() const { return traits_.option_sprite; }
   [[nodiscard]] int OptionOffset(bool focused) const {
     return focused ? traits_.focus_option_offset : traits_.option_offset;
   }
 
-  virtual void FireMain(Player &player, uint8_t tier, bool focused) = 0;
-  virtual void FireSub(Player &player, uint8_t tier, bool focused) = 0;
+  virtual void FireMain(Player &player, int tier, bool focused) = 0;
+  virtual void FireSub(Player &player, int tier, bool focused) = 0;
   virtual void UpdateBomb(Player & /*player*/, EnemyManager & /*enemies*/,
-                          EffectManager & /*effects*/, uint16_t /*remaining*/) {
+                          EffectManager & /*effects*/, int /*remaining*/) {
   }
   virtual void Tick(Player & /*player*/) {}
   virtual void ApplyContinuousAttack(const Player & /*player*/,
                                      EnemyManager & /*enemies*/,
                                      bool /*focused*/) const {}
   virtual void DrawBombBackground(const Player & /*player*/,
-                                  uint16_t /*remaining*/) const {}
+                                  int /*remaining*/) const {}
   virtual void DrawBombForeground(const Player & /*player*/,
-                                  uint16_t /*remaining*/) const {}
+                                  int /*remaining*/) const {}
   virtual void DrawContinuousAttack(const Player & /*player*/,
                                     bool /*focused*/) const {}
   virtual void ClearContinuousAttack() {}
